requirements.txt
logging
datetime
requests
pandas
numpy
PyMySQL
SQLAlchemy
python-dateutil
mysql-connector-python
python-dotenv
onepassword-sdk
cryptography